@@ -1,28 +1,45 @@
 {
-    "name": "fullbowody-website",
-    "version": "0.1.0",
-    "scripts": {
-        "dev": "vite --host --port=80",
-        "build": "vite build",
-        "preview": "vite preview --host --port=80"
-    },
-    "dependencies": {
-        "@heroicons/vue": "^2.0.14",
-        "eslint-plugin-vue": "^9.10.0",
-        "highlight.js": "^11.8.0",
-        "three": "^0.154.0",
-        "vite": "^2.9.15",
-        "vue": "^3.2.25",
-        "vue-router": "^4.0.12"
-    },
-    "devDependencies": {
-        "@vitejs/plugin-vue": "^2.0.0",
-        "postcss": "^8.4.5",
-        "tailwindcss": "^3.0.17"
-    },
-    "author": {
-        "name": "Furwaz",
-        "email": "fur.waz06@gmail.com",
-        "url": "http://furwaz.fr"
-    }
+  "name": "web",
+  "private": true,
+  "version": "1.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite --port 80 --host",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "preview": "vite preview --port 80 --host",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build --force",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
+    "format": "prettier --write src/"
+  },
+  "author": {
+    "name": "FurWaz",
+    "email": "fur.waz06@gmail.com"
+  },
+  "dependencies": {
+    "@heroicons/vue": "^2.1.5",
+    "@types/three": "^0.167.1",
+    "three": "^0.167.1",
+    "vue": "^3.4.35",
+    "vue-router": "^4.4.2"
+  },
+  "devDependencies": {
+    "@rushstack/eslint-patch": "^1.8.0",
+    "@tsconfig/node20": "^20.1.4",
+    "@types/node": "^20.14.5",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^13.0.0",
+    "@vue/tsconfig": "^0.5.1",
+    "autoprefixer": "^10.4.19",
+    "eslint": "^8.57.0",
+    "eslint-plugin-vue": "^9.23.0",
+    "npm-run-all2": "^6.2.0",
+    "postcss": "^8.4.38",
+    "prettier": "^3.2.5",
+    "tailwindcss": "^3.4.4",
+    "typescript": "~5.4.0",
+    "vite": "^5.3.1",
+    "vue-tsc": "^2.0.21"
+  }
 }
