@@ -1,9 +1,27 @@
-module.exports = {
-    content: [
-        "./index.html",
-        "./src/**/*.{vue,js,ts,jsx,tsx}",
+import plugin from 'tailwindcss/plugin'
+
+/** @type {import('tailwindcss').Config} */
+export default {
+    content: ["./index.html", "./src/**/*.{html,js,vue}"],
+    theme: {
+        extend: {
+            transitionTimingFunction: {
+                'in-expo': 'cubic-bezier(0.95, 0.05, 0.795, 0.035)',
+                'out-expo': 'cubic-bezier(0.19, 1, 0.22, 1)',
+            }
+        },
+    },
+    plugins: [
+        plugin(function ({ matchUtilities, theme }) {
+            matchUtilities(
+                {
+                    'translate-z': (value) => ({
+                        '--tw-translate-z': value,
+                        transform: ` translate3d(var(--tw-translate-x), var(--tw-translate-y), var(--tw-translate-z)) rotate(var(--tw-rotate)) skewX(var(--tw-skew-x)) skewY(var(--tw-skew-y)) scaleX(var(--tw-scale-x)) scaleY(var(--tw-scale-y))`,
+                    }), // this is actual CSS
+                },
+                { values: theme('translate'), supportsNegativeValues: true }
+            )
+        })
     ],
-    theme: {},
-    plugins: [],
-    darkMode: "class",
 }
